Move speedy vertical_diffusion to vertical_diffusion/speedy_vdiff.py
Update imports in speedy_physics.py, speedy_terms.py, and test file.
