+++ jcm/physics/vertical_diffusion/speedy_vdiff_test.py
@@ -17,7 +17,7 @@ class Test_VerticalDiffusion_Unit(unittest.TestCase):
         from jcm.physics.speedy.params import Parameters
         from jcm.forcing import ForcingData
         from jcm.physics_interface import PhysicsState, PhysicsTendency
-        from jcm.physics.speedy.vertical_diffusion import get_vertical_diffusion_tend
+        from jcm.physics.vertical_diffusion.speedy_vdiff import get_vertical_diffusion_tend
         from jcm.terrain import TerrainData
         from jcm.physics.speedy.speedy_coords import SpeedyCoords
